Automated cherry pick of #11988: fix(cloudid): hcso custom role

+++ pkg/multicloud/huaweistack/cloudgroup.go
@@ -71,7 +71,7 @@ func (group *SCloudgroup) DetachSystemPolicy(roleId string) error {
 }
 
 func (group *SCloudgroup) DetachCustomPolicy(roleId string) error {
-	return group.client.DetachGroupRole(group.Id, roleId)
+	return group.client.DetachGroupCustomRole(group.Id, roleId)
 }
 
 func (group *SCloudgroup) AttachSystemPolicy(roleId string) error {
@@ -79,7 +79,7 @@ func (group *SCloudgroup) AttachSystemPolicy(roleId string) error {
 }
 
 func (group *SCloudgroup) AttachCustomPolicy(roleId string) error {
-	return group.client.AttachGroupRole(group.Id, roleId)
+	return group.client.AttachGroupCustomRole(group.Id, roleId)
 }
 
 func (group *SCloudgroup) GetISystemCloudpolicies() ([]cloudprovider.ICloudpolicy, error) {
@@ -89,13 +89,25 @@ func (group *SCloudgroup) GetISystemCloudpolicies() ([]cloudprovider.ICloudpolic
 	}
 	ret := []cloudprovider.ICloudpolicy{}
 	for i := range roles {
-		ret = append(ret, &roles[i])
+		if roles[i].Catelog != "CUSTOMED" {
+			ret = append(ret, &roles[i])
+		}
 	}
 	return ret, nil
 }
 
 func (group *SCloudgroup) GetICustomCloudpolicies() ([]cloudprovider.ICloudpolicy, error) {
-	return []cloudprovider.ICloudpolicy{}, nil
+	roles, err := group.client.GetGroupRoles(group.Id)
+	if err != nil {
+		return nil, errors.Wrap(err, "GetGroupRoles")
+	}
+	ret := []cloudprovider.ICloudpolicy{}
+	for i := range roles {
+		if roles[i].Catelog == "CUSTOMED" {
+			ret = append(ret, &roles[i])
+		}
+	}
+	return ret, nil
 }
 
 func (group *SCloudgroup) GetICloudusers() ([]cloudprovider.IClouduser, error) {
@@ -261,21 +273,68 @@ func (self *SHuaweiClient) DetachGroupRole(groupId, roleId string) error {
 	if err != nil {
 		return errors.Wrapf(err, "GetRole(%s)", roleId)
 	}
-	err = client.Groups.DeleteRole(self.ownerId, groupId, role.Id)
-	if err != nil {
-		return errors.Wrapf(err, "DeleteRole")
-	}
-	projects, err := self.GetProjects()
-	if err != nil {
-		return errors.Wrapf(err, "GetProjects")
-	}
-	for _, project := range projects {
-		err = client.Groups.DeleteProjectRole(project.ID, groupId, role.Id)
+	if role.Type == "AX" || role.Type == "AA" {
+		err = client.Groups.DeleteRole(self.ownerId, groupId, role.Id)
 		if err != nil {
-			return errors.Wrapf(err, "DeleteProjectRole")
+			return errors.Wrapf(err, "DeleteRole")
+		}
+	}
+	if role.Type == "XA" || role.Type == "AA" {
+		projects, err := self.GetProjects()
+		if err != nil {
+			return errors.Wrapf(err, "GetProjects")
+		}
+		for _, project := range projects {
+			err = client.Groups.DeleteProjectRole(project.ID, groupId, role.Id)
+			if err != nil {
+				return errors.Wrapf(err, "DeleteProjectRole")
+			}
 		}
 	}
 	return nil
+}
+
+func (self *SHuaweiClient) DetachGroupCustomRole(groupId, roleId string) error {
+	client, err := self.newGeneralAPIClient()
+	if err != nil {
+		return errors.Wrap(err, "newGeneralAPIClient")
+	}
+	role, err := self.GetCustomRole(roleId)
+	if err != nil {
+		return errors.Wrapf(err, "GetCustomRole(%s)", roleId)
+	}
+	if role.Type == "AX" || role.Type == "AA" {
+		err = client.Groups.DeleteRole(self.ownerId, groupId, role.Id)
+		if err != nil {
+			return errors.Wrapf(err, "DeleteRole")
+		}
+	}
+	if role.Type == "XA" || role.Type == "AA" {
+		projects, err := self.GetProjects()
+		if err != nil {
+			return errors.Wrapf(err, "GetProjects")
+		}
+		for _, project := range projects {
+			err = client.Groups.DeleteProjectRole(project.ID, groupId, role.Id)
+			if err != nil {
+				return errors.Wrapf(err, "DeleteProjectRole")
+			}
+		}
+	}
+	return nil
+}
+
+func (self *SHuaweiClient) GetCustomRole(name string) (*SRole, error) {
+	roles, err := self.GetCustomRoles()
+	if err != nil {
+		return nil, errors.Wrapf(err, "GetCustomRoles(%s)", name)
+	}
+	for i := range roles {
+		if roles[i].DisplayName == name {
+			return &roles[i], nil
+		}
+	}
+	return nil, errors.Wrapf(cloudprovider.ErrNotFound, name)
 }
 
 func (self *SHuaweiClient) GetRole(name string) (*SRole, error) {
@@ -300,18 +359,52 @@ func (self *SHuaweiClient) AttachGroupRole(groupId, roleId string) error {
 	if err != nil {
 		return errors.Wrapf(err, "GetRole(%s)", roleId)
 	}
-	err = client.Groups.AddRole(self.ownerId, groupId, role.Id)
-	if err != nil {
-		return errors.Wrapf(err, "AddRole")
-	}
-	projects, err := self.GetProjects()
-	if err != nil {
-		return errors.Wrapf(err, "GetProjects")
-	}
-	for _, project := range projects {
-		err = client.Groups.AddProjectRole(project.ID, groupId, role.Id)
+	if role.Type == "AX" || role.Type == "AA" {
+		err = client.Groups.AddRole(self.ownerId, groupId, role.Id)
 		if err != nil {
-			return errors.Wrapf(err, "AddProjectRole")
+			return errors.Wrapf(err, "AddRole")
+		}
+	}
+	if role.Type == "XA" || role.Type == "AA" {
+		projects, err := self.GetProjects()
+		if err != nil {
+			return errors.Wrapf(err, "GetProjects")
+		}
+		for _, project := range projects {
+			err = client.Groups.AddProjectRole(project.ID, groupId, role.Id)
+			if err != nil {
+				return errors.Wrapf(err, "AddProjectRole")
+			}
+		}
+	}
+	return nil
+}
+
+func (self *SHuaweiClient) AttachGroupCustomRole(groupId, roleId string) error {
+	client, err := self.newGeneralAPIClient()
+	if err != nil {
+		return errors.Wrap(err, "newGeneralAPIClient")
+	}
+	role, err := self.GetCustomRole(roleId)
+	if err != nil {
+		return errors.Wrapf(err, "GetRole(%s)", roleId)
+	}
+	if role.Type == "AX" || role.Type == "AA" {
+		err = client.Groups.AddRole(self.ownerId, groupId, role.Id)
+		if err != nil {
+			return errors.Wrapf(err, "AddRole")
+		}
+	}
+	if role.Type == "XA" || role.Type == "AA" {
+		projects, err := self.GetProjects()
+		if err != nil {
+			return errors.Wrapf(err, "GetProjects")
+		}
+		for _, project := range projects {
+			err = client.Groups.AddProjectRole(project.ID, groupId, role.Id)
+			if err != nil {
+				return errors.Wrapf(err, "AddProjectRole")
+			}
 		}
 	}
 	return nil
